stable commit
Update:
Chapter 3: The unit tests for the derivative functions are golden. The update functions not so much. Dk what the problem is.
Chapter 4: wind is golden. Torque and thrust not working (syntax error being bitchy). For the forces and moments, My, Mz, Fy are working perfectly. The other 3 not. Found a typo that fixed fy, still searching for the others.